mb fixed freeze in find_intersection kernel (removed call of old render when pathtrace is on) + fixed incorrect data interpretation (kernel work sizes changed from size_t to uint32_t to match them on device). works fine on GTX 1060, todo test on school's Macs with Radeon

+++ includes/rt_opencl.h
@@ -5,9 +5,9 @@
 
 typedef struct		s_kernel_work_sizes
 {
-	size_t			materials;
-	size_t			textures;
-	size_t			skybox;
+	uint32_t		materials;
+	uint32_t		textures;
+	uint32_t		skybox;
 }					t_kernel_work_sizes;
 
 void			rt_opencl_prepare_memory(t_rt *rt, t_rt_renderer *renderer);
